+++ core/components/romanescobackyard/lexicon/en/default.inc.php
@@ -1,17 +1,13 @@
 <?php
 
-// WEB CONTEXT
-// =====================================================================
-
-// Common website elements
+// Web context
 // ---------------------------------------------------------------------
 
+// Common website elements
 $_lang['romanesco.footer.qredits_text'] = 'A Romanesco project';
 $_lang['romanesco.footer.qredits_title'] = 'Modular websites built with MODX';
 
 // Navigation
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.menu.brand'] = 'Home';
 $_lang['romanesco.menu.overview'] = 'Overview';
 $_lang['romanesco.menu.logo_title'] = 'Homepage of [[++site_name]]';
@@ -24,38 +20,33 @@ $_lang['romanesco.pagination.last'] = 'Last page';
 $_lang['romanesco.pagination.load_more'] = 'Load more';
 
 // Social
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.social.connect_facebook_organization'] = 'Visit our Facebook page';
-$_lang['romanesco.social.connect_googleplus_organization'] = 'Find us on Google+';
-$_lang['romanesco.social.connect_linkedin_organization'] = 'Connect with us on LinkedIn';
 $_lang['romanesco.social.connect_instagram_organization'] = 'Follow us on Instagram';
+$_lang['romanesco.social.connect_linkedin_organization'] = 'Connect with us on LinkedIn';
 $_lang['romanesco.social.connect_pinterest_organization'] = 'View our Pinterest board';
 $_lang['romanesco.social.connect_twitter_organization'] = 'Follow us on Twitter';
 $_lang['romanesco.social.connect_youtube_organization'] = 'View our Youtube channel';
 $_lang['romanesco.social.connect_github_organization'] = 'View our Github profile';
 
-$_lang['romanesco.social.connect_facebook_person'] = 'Visit my Facebook page';
-$_lang['romanesco.social.connect_googleplus_person'] = 'Find me on Google+';
-$_lang['romanesco.social.connect_linkedin_person'] = 'Connect with me on LinkedIn';
+$_lang['romanesco.social.connect_facebook_person'] = 'Find me on Facebook';
 $_lang['romanesco.social.connect_instagram_person'] = 'Follow me on Instagram';
+$_lang['romanesco.social.connect_linkedin_person'] = 'Connect with me on LinkedIn';
 $_lang['romanesco.social.connect_pinterest_person'] = 'View my Pinterest board';
 $_lang['romanesco.social.connect_twitter_person'] = 'Follow me on Twitter';
 $_lang['romanesco.social.connect_youtube_person'] = 'View my Youtube channel';
 $_lang['romanesco.social.connect_github_person'] = 'View my Github profile';
 
-$_lang['romanesco.social.share_facebook_title'] = 'Share on Facebook';
-$_lang['romanesco.social.share_googleplus_title'] = 'Give this a +1';
-$_lang['romanesco.social.share_linkedin_title'] = 'Share on LinkedIn';
-$_lang['romanesco.social.share_pinterest_title'] = 'Pin this';
-$_lang['romanesco.social.share_twitter_title'] = 'Tweet about this';
+$_lang['romanesco.social.share_facebook'] = 'Share on Facebook';
+$_lang['romanesco.social.share_linkedin'] = 'Share on LinkedIn';
+$_lang['romanesco.social.share_pinterest'] = 'Pin this';
+$_lang['romanesco.social.share_twitter'] = 'Tweet about this';
+$_lang['romanesco.social.share_email'] = 'Send link via email';
+$_lang['romanesco.social.share_email_body'] = 'Hi friend, I thought you might find this interesting too:';
 
 // Publication
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.article.written_by'] = 'Written by';
 $_lang['romanesco.article.written_on'] = 'on';
-$_lang['romanesco.article.about_author'] = 'About the author,';
+$_lang['romanesco.article.about_author'] = 'About the author';
 $_lang['romanesco.article.comments'] = 'Comments';
 $_lang['romanesco.article.share_this'] = 'Share this article';
 $_lang['romanesco.article.share_this_short'] = 'Share';
@@ -67,24 +58,18 @@ $_lang['romanesco.article.most_viewed'] = 'Most viewed';
 $_lang['romanesco.article.related'] = 'Related';
 
 // Testimonials
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.testimonials.logo_title'] = 'Read why [[+pagetitle]] loves to work with us!';
 
 // Search
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.search.field_placeholder'] = 'What are you looking for?';
 $_lang['romanesco.search.field_button_text'] = 'Search';
 $_lang['romanesco.search.result_button_text'] = 'Go to this page';
 
 
-// PROJECT HUB CONTEXT
-// =====================================================================
-
-// Status grid
+// Project hub
 // ---------------------------------------------------------------------
 
+// Status grid
 $_lang['romanesco.status.table_indicator'] = 'Status';
 $_lang['romanesco.status.table_id'] = 'ID';
 $_lang['romanesco.status.table_resource'] = 'Page';
@@ -98,10 +83,7 @@ $_lang['romanesco.status.legend'] = 'Legend';
 $_lang['romanesco.status.legend_progress'] = 'Progress indicators';
 $_lang['romanesco.status.legend_health'] = 'Health indicators';
 
-
 // Pattern library
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.patterns.button_backyard'] = 'Backyard';
 $_lang['romanesco.patterns.preview_tooltip'] = 'Hide example';
 $_lang['romanesco.patterns.preview_not_available'] = 'No preview available';
@@ -124,6 +106,7 @@ $_lang['romanesco.patterns.referring_pages'] = 'Used on the following pages';
 $_lang['romanesco.patterns.assigned_templates'] = 'Assigned to the following templates';
 $_lang['romanesco.patterns.assigned_tvs'] = 'Has the following TVs assigned';
 
+// Status indicator
 $_lang['romanesco.patterns.status_green'] = 'Solid (safe to use as instructed)';
 $_lang['romanesco.patterns.status_yellow'] = 'Incoherent (works without issues, but might be refactored at some point)';
 $_lang['romanesco.patterns.status_orange'] = 'Conflicting (there are open issues related to this pattern)';
@@ -132,35 +115,3 @@ $_lang['romanesco.patterns.status_pink'] = 'Experimental (just fiddling with som
 $_lang['romanesco.patterns.status_purple'] = 'Untested (use at your own risk)';
 $_lang['romanesco.patterns.status_blue'] = 'In review (under evaluation if this belongs in the library or not)';
 $_lang['romanesco.patterns.status_brown'] = 'Deprecated (will be deleted in the next major release)';
-
-
-// MANAGER CONTEXT
-// =====================================================================
-
-// Top menu
-// ---------------------------------------------------------------------
-
-//// Global content
-//$_lang['romanesco.menu.global_content_title'] = 'Global Content';
-//$_lang['romanesco.menu.global_content_description'] = 'Change content that\'s available site-wide';
-//
-//$_lang['romanesco.menu.edit_footer_title'] = 'Edit footer';
-//$_lang['romanesco.menu.edit_footer_description'] = 'Change footer content for all pages';
-//
-//$_lang['romanesco.menu.edit_forms_title'] = 'FormBlocks';
-//$_lang['romanesco.menu.edit_forms_description'] = 'Build and edit your own forms';
-//
-//$_lang['romanesco.menu.edit_cta_title'] = 'Call To Actions';
-//$_lang['romanesco.menu.edit_cta_description'] = 'Build and edit your own CTA\'s';
-//
-//// Custom caching
-//$_lang['romanesco.menu.clear_cache_all_title'] = 'Clear Custom Cache';
-//$_lang['romanesco.menu.clear_cache_all_description'] = 'Delete files inside all custom cache partitions';
-//$_lang['romanesco.menu.clear_cache_overviews_title'] = 'Overviews';
-//$_lang['romanesco.menu.clear_cache_overviews_description'] = 'Only clear cache of standard Overview grids';
-//$_lang['romanesco.menu.clear_cache_clients_title'] = 'Clients';
-//$_lang['romanesco.menu.clear_cache_gallery_title'] = 'Gallery';
-//$_lang['romanesco.menu.clear_cache_portfolio_title'] = 'Portfolio';
-//$_lang['romanesco.menu.clear_cache_publication_title'] = 'Publication';
-//$_lang['romanesco.menu.clear_cache_team_title'] = 'Team';
-//$_lang['romanesco.menu.clear_cache_testimonials_title'] = 'Testimonials';