+++ core/components/romanescobackyard/lexicon/nl/default.inc.php
@@ -1,14 +1,13 @@
 <?php
 
-// Common website elements
+// Web context
 // ---------------------------------------------------------------------
 
+// Common website elements
 $_lang['romanesco.footer.qredits_text'] = 'Een Romanesco site';
 $_lang['romanesco.footer.qredits_title'] = 'Modulaire web projecten met MODX CMS';
 
 // Navigation
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.menu.brand'] = 'Home';
 $_lang['romanesco.menu.overview'] = 'Overzicht';
 $_lang['romanesco.menu.logo_title'] = 'Homepage van [[++site_name]]';
@@ -21,35 +20,33 @@ $_lang['romanesco.pagination.last'] = 'Laatste pagina';
 $_lang['romanesco.pagination.load_more'] = 'Laad meer';
 
 // Social
-// ---------------------------------------------------------------------
+$_lang['romanesco.social.connect_facebook_organization'] = 'Bezoek onze Facebook pagina';
+$_lang['romanesco.social.connect_instagram_organization'] = 'Volg ons op Instagram';
+$_lang['romanesco.social.connect_linkedin_organization'] = 'Verbinden via LinkedIn';
+$_lang['romanesco.social.connect_pinterest_organization'] = 'Bekijk ons Pinterest board';
+$_lang['romanesco.social.connect_twitter_organization'] = 'Volg ons op Twitter';
+$_lang['romanesco.social.connect_youtube_organization'] = 'Bekijk ons Youtube kanaal';
+$_lang['romanesco.social.connect_github_organization'] = 'Bekijk ons Github profiel';
 
-$_lang['romanesco.social.connect_facebook_title'] = 'Bezoek onze Facebook pagina';
-$_lang['romanesco.social.connect_googleplus_title'] = 'Bekijk ons Google+ profiel';
-$_lang['romanesco.social.connect_instagram_title'] = 'Volg ons op Instagram';
-$_lang['romanesco.social.connect_linkedin_title'] = 'Verbinden via LinkedIn';
-$_lang['romanesco.social.connect_pinterest_title'] = 'Bekijk ons Pinterest board';
-$_lang['romanesco.social.connect_twitter_title'] = 'Volg ons op Twitter';
-$_lang['romanesco.social.connect_youtube_title'] = 'Bekijk ons Youtube kanaal';
-
-$_lang['romanesco.social.connect_facebook_person'] = 'Facebook';
-$_lang['romanesco.social.connect_googleplus_person'] = 'Google+';
-$_lang['romanesco.social.connect_linkedin_person'] = 'LinkedIn';
-$_lang['romanesco.social.connect_instagram_person'] = 'Instagram';
-$_lang['romanesco.social.connect_pinterest_person'] = 'Pinterest';
-$_lang['romanesco.social.connect_twitter_person'] = 'Twitter';
-$_lang['romanesco.social.connect_youtube_person'] = 'Youtube';
+$_lang['romanesco.social.connect_facebook_person'] = 'Bezoek mijn Facebook pagina';
+$_lang['romanesco.social.connect_instagram_person'] = 'Volg mij op Instagram';
+$_lang['romanesco.social.connect_linkedin_person'] = 'Verbinden via LinkedIn';
+$_lang['romanesco.social.connect_pinterest_person'] = 'Bekijk mijn Pinterest board';
+$_lang['romanesco.social.connect_twitter_person'] = 'Volg mij op Twitter';
+$_lang['romanesco.social.connect_youtube_person'] = 'Bekijk mijn Youtube kanaal';
+$_lang['romanesco.social.connect_github_person'] = 'Bekijk mijn Github profiel';
 
 $_lang['romanesco.social.share_facebook_title'] = 'Delen via Facebook';
-$_lang['romanesco.social.share_googleplus_title'] = 'Geef dit een +1';
 $_lang['romanesco.social.share_linkedin_title'] = 'Delen via LinkedIn';
 $_lang['romanesco.social.share_pinterest_title'] = 'Pinnen';
 $_lang['romanesco.social.share_twitter_title'] = 'Tweeten';
+$_lang['romanesco.social.share_email'] = 'Delen via email';
+$_lang['romanesco.social.share_email_body'] = 'Hoi vriend, volgens mij vind jij dit ook wel interessant:';
 
 // Publication
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.article.written_by'] = 'Geschreven door';
 $_lang['romanesco.article.written_on'] = 'op';
+$_lang['romanesco.article.about_author'] = 'Over de auteur';
 $_lang['romanesco.article.comments'] = 'Reacties';
 $_lang['romanesco.article.share_this'] = 'Deel dit artikel';
 $_lang['romanesco.article.share_this_short'] = 'Delen';
@@ -60,40 +57,25 @@ $_lang['romanesco.article.latest'] = 'Laatst toegevoegd';
 $_lang['romanesco.article.most_viewed'] = 'Meest gelezen';
 $_lang['romanesco.article.related'] = 'Gerelateerde artikelen';
 
-
 // Testimonials
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.testimonials.logo_title'] = 'Lees waarom [[+pagetitle]] graag met ons samenwerkt!';
 
 // Search
-// ---------------------------------------------------------------------
-
 $_lang['romanesco.search.field_placeholder'] = 'Waar ben je naar op zoek?';
 $_lang['romanesco.search.field_button_text'] = 'Zoeken';
 $_lang['romanesco.search.result_button_text'] = 'Ga naar deze pagina';
 
 
-// PROJECT HUB CONTEXT
-// =====================================================================
-
-// Status grid
+// Project hub
 // ---------------------------------------------------------------------
 
-$_lang['romanesco.status.table_status_indicator'] = 'Status';
-$_lang['romanesco.status.table_id'] = 'ID';
-$_lang['romanesco.status.table_resource'] = 'Pagina';
-$_lang['romanesco.status.table_improvements'] = 'Verbeteringen';
-$_lang['romanesco.status.table_assigned_user'] = 'Beheerder';
-$_lang['romanesco.status.table_help'] = 'Problemen';
-$_lang['romanesco.status.table_notes'] = 'Notities / Toelichting';
-
+// Status grid
 $_lang['romanesco.status.table_indicator'] = 'Status';
 $_lang['romanesco.status.table_id'] = 'ID';
 $_lang['romanesco.status.table_resource'] = 'Pagina';
 $_lang['romanesco.status.table_improvements'] = 'Verbeteringen';
 $_lang['romanesco.status.table_issues'] = 'Problemen';
-$_lang['romanesco.status.table_indexation'] = 'Indexatie';
+$_lang['romanesco.status.table_indexation'] = 'Indexering';
 $_lang['romanesco.status.table_owner'] = 'Beheerder';
 $_lang['romanesco.status.table_date_due'] = 'Vervaldatum';